feat: separate community and address pages
