+++ components/CommunityPageContent/CommunityPageContent.module.css
@@ -1,0 +1,8 @@
+.container {
+  width: '100%';
+  height: '100%';
+  display: 'flex';
+  flex-direction: 'column';
+  align-items: 'center';
+  flex-grow: 1;
+}
