+++ components/CommunityPageContent/index.ts
@@ -1,0 +1,1 @@
+export { CommunityAddressPage, CommunityPage } from './CommunityPageContent';
